refactor(money): Faz 3 double → Decimal para tutarı migration
CLAUDE.md "Yasak Listesi" sözleşmesi: para için double/float YASAK.
IEEE-754 binary representation 0.1 + 0.2 != 0.3 üretiyor; finansal
toplamada kullanıcı 1 kuruşluk fark görüyor → güven kaybı. Bu commit 5
feature'daki para alanlarını Decimal'a geçiriyor.

## Yeni paket
- decimal: ^3.0.2 (pubspec.yaml + lock)

## Yeni helper
- lib/core/utils/money_parser.dart: MoneyParser
  - requireDecimal(json, field) → zorunlu para alanı, FormatException
  - tryDecimal(value) → opsiyonel, null tolere, num/double/String hepsi kabul
  - toJsonString(decimal) → backend'e String olarak serialize
- test/core/utils/money_parser_test.dart: 15 case (null, int, double,
  String, Decimal, boş, NaN, Infinity, invalid, bool, FormatException,
  0.1+0.2=0.3 invariant, finansal toplama)

## Domain entity değişiklikleri (para → Decimal, yüzde → double)
- WhatIfResult: buyPrice, sellPrice, unitsAcquired, initialValueTry,
  finalValueTry, profitLossTry → Decimal. profitLossPercent, *Percent →
  double (display-only)
- ChartPoint.price → Decimal
- ReverseWhatIfResult: tüm para alanları → Decimal
- DcaResult: periodicAmount, totalInvestedTry, currentValueTry,
  profitLossTry, averageCostPerUnit, totalUnitsAcquired,
  currentUnitPrice → Decimal
- DcaPurchase, DcaChartPoint: price, units, cost, value alanları → Decimal
- PortfolioResult: totalInitial/Final/ProfitLossTry, totalRealProfitLossTry
  → Decimal
- SavedScenario.amount: num → Decimal (F-11-02)

## Veri katmanı (JSON parse)
- what_if_response_model.dart, reverse_what_if_response_model.dart,
  dca_response_model.dart, saved_scenario_model.dart: para alanları
  MoneyParser.requireDecimal ile parse. Backend hem string hem num
  döndürse de güvenle çevriliyor.

## Use case
- calculate_portfolio.dart: aggregasyon artık Decimal aritmetiğinde
  (totalInitial, totalFinal, totalPnL toplamı 0.1+0.2 hatası yapmaz).
  Yüzde hesapları Decimal/Decimal Rational döndüğü için .toDouble()
  ile cast.

## Sunum katmanı (.toDouble() boundary)
- WhatIf, DCA, Portfolio, Comparison ResultCard / ShareCard'larında
  NumberFormat.currency.format(decimal.toDouble()) — precision loss
  display'de tolere edilir (zaten 2 ondalık haneye yuvarlar).
- result_chart, dca_chart: FlSpot.y double bekliyor — .toDouble().
- scenario_card.dart: scenario.amount.toDouble() format'a sokuluyor.

## Event/BLoC boundary
- ScenarioSaveRequested.amount num olarak korunuyor (user input
  boundary, double-exact). app.dart ve portfolio_page.dart
  scenario.amount.toDouble() ile event'e geçiriyor.
- ScenariosBloc duplicate check `s.amount.toDouble() == event.amount`.

## Test güncellemeleri
- what_if_response_model_test.dart: 5.95 → Decimal.parse('5.95')
- scenarios_bloc_test.dart: amount: 10000 → Decimal.fromInt(10000)
- Hiç test silinmedi; 146 test yeşil (önceki 126 + Faz 2.1/2.4 yeni
  testler + MoneyParser 15 case).

## Doc
- docs/architecture.md "Sonuç Gösterimi Kuralları" altına yeni alt
  başlık: "Para Tutarı için Decimal" — kural + örnek + Decimal/Rational
  notu eklendi.

`flutter analyze --fatal-infos`: temiz. `flutter test`: 146 passed.

+++ lib/app.dart
@@ -172,7 +172,10 @@ class _MainShellState extends State<MainShell> {
             assetSymbol: scenario.assetSymbol,
             buyDate: scenario.buyDate,
             sellDate: scenario.sellDate,
-            amount: scenario.amount,
+            // Replay event num bekliyor (form input num); kullanıcı
+            // tarafından girilmiş tutar zaten double-exact (örn. 47010.34)
+            // — Decimal → double burada precision farkı yaratmaz.
+            amount: scenario.amount.toDouble(),
             amountType: scenario.amountType,
             includeInflation:
                 (scenario.extraData?['includeInflation'] as bool?) ?? false,
@@ -188,7 +191,7 @@ class _MainShellState extends State<MainShell> {
             symbols: scenario.assetSymbol.split(','),
             buyDate: scenario.buyDate,
             sellDate: scenario.sellDate,
-            amount: scenario.amount,
+            amount: scenario.amount.toDouble(),
             includeInflation:
                 (scenario.extraData?['includeInflation'] as bool?) ?? false,
           ),
@@ -246,7 +249,8 @@ class _MainShellState extends State<MainShell> {
             startDate: scenario.buyDate,
             endDate: scenario.sellDate,
             periodicAmount:
-                (extra?['periodicAmount'] as num?) ?? scenario.amount,
+                (extra?['periodicAmount'] as num?) ??
+                scenario.amount.toDouble(),
             period: (extra?['period'] as String?) ?? 'monthly',
             amountType: scenario.amountType,
             includeInflation: (extra?['includeInflation'] as bool?) ?? false,
